refactor(app): extract PermissionRow to its own file
Reduces SettingsView.swift below 600-line SwiftLint limit.
